added core properties opt

+++ contracts/core/CoreProperties.sol
@@ -66,11 +66,11 @@ contract CoreProperties is ICoreProperties, OwnableUpgradeable, AbstractDependan
         _blacklistTokens.remove(tokens);
     }
 
-    function setMaximumPoolInvestors(uint256 count) external override onlyOwner {
+    function setMaximumPoolInvestors(uint64 count) external override onlyOwner {
         coreParameters.traderParams.maxPoolInvestors = count;
     }
 
-    function setMaximumOpenPositions(uint256 count) external override onlyOwner {
+    function setMaximumOpenPositions(uint64 count) external override onlyOwner {
         coreParameters.traderParams.maxOpenPositions = count;
     }
 
@@ -82,11 +82,11 @@ contract CoreProperties is ICoreProperties, OwnableUpgradeable, AbstractDependan
         coreParameters.traderParams.leverageSlope = slope;
     }
 
-    function setCommissionInitTimestamp(uint256 timestamp) external override onlyOwner {
+    function setCommissionInitTimestamp(uint64 timestamp) external override onlyOwner {
         coreParameters.traderParams.commissionInitTimestamp = timestamp;
     }
 
-    function setCommissionDurations(uint256[] calldata durations) external override onlyOwner {
+    function setCommissionDurations(uint64[] calldata durations) external override onlyOwner {
         coreParameters.traderParams.commissionDurations = durations;
     }
 
@@ -110,7 +110,7 @@ contract CoreProperties is ICoreProperties, OwnableUpgradeable, AbstractDependan
         coreParameters.traderParams.maxTraderCommissions = maxTraderCommissions;
     }
 
-    function setDelayForRiskyPool(uint256 delayForRiskyPool) external override onlyOwner {
+    function setDelayForRiskyPool(uint64 delayForRiskyPool) external override onlyOwner {
         coreParameters.traderParams.delayForRiskyPool = delayForRiskyPool;
     }
 
@@ -176,11 +176,11 @@ contract CoreProperties is ICoreProperties, OwnableUpgradeable, AbstractDependan
         }
     }
 
-    function getMaximumPoolInvestors() external view override returns (uint256) {
+    function getMaximumPoolInvestors() external view override returns (uint64) {
         return coreParameters.traderParams.maxPoolInvestors;
     }
 
-    function getMaximumOpenPositions() external view override returns (uint256) {
+    function getMaximumOpenPositions() external view override returns (uint64) {
         return coreParameters.traderParams.maxOpenPositions;
     }
 
@@ -191,13 +191,11 @@ contract CoreProperties is ICoreProperties, OwnableUpgradeable, AbstractDependan
         );
     }
 
-    function getCommissionInitTimestamp() public view override returns (uint256) {
+    function getCommissionInitTimestamp() public view override returns (uint64) {
         return coreParameters.traderParams.commissionInitTimestamp;
     }
 
-    function getCommissionDuration(
-        CommissionPeriod period
-    ) public view override returns (uint256) {
+    function getCommissionDuration(CommissionPeriod period) public view override returns (uint64) {
         return coreParameters.traderParams.commissionDurations[uint256(period)];
     }
 
@@ -222,7 +220,7 @@ contract CoreProperties is ICoreProperties, OwnableUpgradeable, AbstractDependan
         );
     }
 
-    function getDelayForRiskyPool() external view override returns (uint256) {
+    function getDelayForRiskyPool() external view override returns (uint64) {
         return coreParameters.traderParams.delayForRiskyPool;
     }
